First pass at merging webpack configurations
