+++ .github/workflows/build.yaml
@@ -17,7 +17,6 @@ jobs:
       - uses: bahmutov/npm-install@v1
       - run: npm run build:chrome
         env:
-          NODE_ENV: production
           ENVIRONMENT: production
           ROLLBAR_PUBLIC_PATH: chrome-extension://production/
           CHROME_EXTENSION_ID: mpjjildhmpddojocokjkgmlkkkfjnepo
@@ -47,7 +46,6 @@ jobs:
       - uses: bahmutov/npm-install@v1
       - run: npm run build:chrome
         env:
-          NODE_ENV: production
           ENVIRONMENT: production
           ROLLBAR_PUBLIC_PATH: chrome-extension://production/
           CHROME_EXTENSION_ID: mpjjildhmpddojocokjkgmlkkkfjnepo
@@ -84,7 +82,6 @@ jobs:
       - uses: bahmutov/npm-install@v1
       - run: npm run build:firefox
         env:
-          NODE_ENV: production
           ENVIRONMENT: production
           ROLLBAR_PUBLIC_PATH: moz-extension://production/
           SERVICE_URL: https://app.pixiebrix.com
