--- browsers/firefox/manifests/manifest.template.json
@@ -1,50 +1,0 @@
-{
-  "manifest_version": 2,
-  "author": "Pixie Brix, LLC",
-  "name": "PixieBrix",
-  "short_name": "PixieBrix",
-  "description": "PixieBrix - integrate your favorite web applications without code",
-  "icons": {
-    "16": "icons/logo16.png",
-    "32": "icons/logo32.png",
-    "48": "icons/logo48.png",
-    "128": "icons/logo128.png"
-  },
-  "content_security_policy": "default-src 'self'",
-  "content_scripts": [
-    {
-      "matches": ["https://*.pixiebrix.com/*"],
-      "js": ["contentScript.js"],
-      "css": ["styles.css"],
-      "run_at": "document_idle"
-    }
-  ],
-  "optional_permissions": [
-    "notifications",
-    "clipboardWrite",
-    "https://*/",
-    "http://*/"
-  ],
-  "permissions": [
-    "storage",
-    "identity",
-    "tabs",
-    "webNavigation",
-    "contextMenus"
-  ],
-  "web_accessible_resources": [
-    "css/*",
-    "bundles/*",
-    "frame.html",
-    "frame.css",
-    "script.js",
-    "audio/*"
-  ],
-  "background": {
-    "page": "background.html"
-  },
-  "options_ui": {
-    "page": "options.html",
-    "open_in_tab": true
-  }
-}
